Replace ImageTools with extension methods.
